move class Point (from base.h -> point.h)
