+++ lib/target.h
@@ -1,6 +1,7 @@
 #pragma once
 
 #include "base.h"
+#include "point.h"
 
 namespace lib {
